将部分功能抽离Device

+++ KRender/Interface/IKSwapChain.h
@@ -4,9 +4,9 @@
 struct IKSwapChain
 {
 	virtual ~IKSwapChain() {}
-	virtual bool Init(uint32_t width, uint32_t height, size_t frameInFlight) = 0;
+	virtual bool Init(uint32_t width, uint32_t height, uint32_t frameInFlight) = 0;
 	virtual bool UnInit() = 0;
 	virtual uint32_t GetWidth() = 0;
 	virtual uint32_t GetHeight() = 0;
-	virtual IKRenderTargetPtr GetRenderTarget(size_t frameIndex) = 0;
+	virtual IKRenderTargetPtr GetRenderTarget(uint32_t frameIndex) = 0;
 };
